requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
python-multipart>=0.0.9
SQLAlchemy>=2.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
PyYAML>=6.0
APScheduler>=3.10
itsdangerous>=2.1
